Rework auth guard on mobile
