+++ front/apps/mobile/app/(app)/browse/index.tsx
