+++ front/apps/mobile/app/(app)/collection/[slug]/index.tsx
