+++ front/apps/mobile/app/(app)/movie/[slug]/index.tsx
