+++ front/apps/mobile/app/(app)/movie/[slug]/watch.tsx
