+++ front/apps/mobile/app/(app)/search/index.tsx
