+++ front/apps/mobile/app/(app)/show/[slug].tsx
