+++ front/apps/mobile/app/(app)/watch/[slug].tsx
